Linear search

+++ 0x1E-search_algorithms/0-linear.c
@@ -14,7 +14,7 @@ int linear_search(int *array, size_t size, int value);
 {
 	if (array == NULL)
 		return (-1);
-	for (int i = 0; i < size; i++)
+	for (int i = 0; i < (int) size; i++)
 	{
 		printf("Value checked array[%u] = [%d]\n", i, array[i]);
 		if (value == array[i])
